Extract annotated test code to utilities crate
