+++ utilities/src/lib.rs
@@ -1,2 +1,3 @@
+pub mod annotated_test;
 pub mod bench;
 pub mod test_program;
